+++ src/python/arcor2_mocks/scripts/mock_scene.py
@@ -8,6 +8,7 @@ from typing import Dict
 import humps
 from flask import jsonify, request
 
+from arcor2 import env
 from arcor2.data import common, object_type, scene
 from arcor2.flask import RespT, create_app, run_app
 from arcor2_mocks import SCENE_PORT, SCENE_SERVICE_NAME, version
@@ -16,6 +17,18 @@ app = create_app(__name__)
 
 collision_objects: Dict[str, object_type.Models] = {}
 started: bool = False
+
+delay_mean = env.get_float("ARCOR2_MOCK_SCENE_DELAY_MEAN", 0)
+delay_sigma = env.get_float("ARCOR2_MOCK_SCENE_DELAY_SIGMA", 0)
+
+
+def delay() -> None:
+    """This is to simulate the long-starting Scene service.
+
+    Could be useful to uncover specific kind of bugs.
+    :return:
+    """
+    time.sleep(random.normalvariate(delay_mean, delay_sigma))
 
 
 @app.route("/collisions/box", methods=["PUT"])
@@ -64,7 +77,7 @@ def put_box() -> RespT:
     box = object_type.Box(args["boxId"], float(args["sizeX"]), float(args["sizeY"]), float(args["sizeZ"]))
 
     collision_objects[box.id] = box
-    return "ok", 200
+    return jsonify("ok"), 200
 
 
 @app.route("/collisions/sphere", methods=["PUT"])
@@ -98,7 +111,7 @@ def put_sphere() -> RespT:
     args = humps.decamelize(request.args.to_dict())
     sphere = object_type.Sphere(args["sphere_id"], float(args["radius"]))
     collision_objects[sphere.id] = sphere
-    return "ok", 200
+    return jsonify("ok"), 200
 
 
 @app.route("/collisions/cylinder", methods=["PUT"])
@@ -137,7 +150,7 @@ def put_cylinder() -> RespT:
     args = humps.decamelize(request.args.to_dict())
     cylinder = object_type.Cylinder(args["cylinder_id"], float(args["radius"]), float(args["height"]))
     collision_objects[cylinder.id] = cylinder
-    return "ok", 200
+    return jsonify("ok"), 200
 
 
 @app.route("/collisions/mesh", methods=["PUT"])
@@ -188,7 +201,7 @@ def put_mesh() -> RespT:
     args = humps.decamelize(request.args.to_dict())
     mesh = object_type.Mesh(args["mesh_id"], args["uri"])
     collision_objects[mesh.id] = mesh
-    return "ok", 200
+    return jsonify("ok"), 200
 
 
 @app.route("/collisions/<string:collisionId>", methods=["DELETE"])
@@ -214,9 +227,9 @@ def delete_collision(collisionId: str) -> RespT:
     try:
         del collision_objects[collisionId]
     except KeyError:
-        return "Not found", 404
+        return jsonify("Not found"), 404
 
-    return "ok", 200
+    return jsonify("ok"), 200
 
 
 @app.route("/collisions", methods=["GET"])
@@ -268,22 +281,67 @@ def put_focus() -> RespT:
 
 @app.route("/system/start", methods=["PUT"])
 def put_start() -> RespT:
+    """Starts the system.
+    ---
+    put:
+        tags:
+            - System
+        description: Starts the system.
+        responses:
+            200:
+              description: Ok
+              content:
+                  application/json:
+                      schema:
+                        type: string
+    """
+
     global started
-    time.sleep(random.uniform(0.5, 5.0))
+    delay()
     started = True
-    return "ok", 200
+    return jsonify("ok"), 200
 
 
 @app.route("/system/stop", methods=["PUT"])
 def put_stop() -> RespT:
+    """Stops the system.
+    ---
+    put:
+        tags:
+            - System
+        description: Stops the system.
+        responses:
+            200:
+              description: Ok
+              content:
+                  application/json:
+                      schema:
+                        type: string
+    """
+
     global started
-    time.sleep(random.uniform(0.5, 5.0))
+    delay()
     started = False
-    return "ok", 200
+    return jsonify("ok"), 200
 
 
 @app.route("/system/running", methods=["GET"])
 def get_started() -> RespT:
+    """Gets system state.
+    ---
+    get:
+        tags:
+            - System
+        description: Gets system state.
+        responses:
+            200:
+              description: Ok
+              content:
+                  application/json:
+                      schema:
+                        type: boolean
+    """
+
     return jsonify(started)