+++ src/python/arcor2_mocks/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [0.13.0] - 2021-04-20
+
+### Fixed
+- Methods from System controller missed OpenAPI definition.
+
+### Changed
+- Configurable delay on start/stop.
+  - Scene mock can simulate that start and stop take some time.
+  - Mean time can be set using `ARCOR2_MOCK_SCENE_DELAY_MEAN`.
+  - `ARCOR2_MOCK_SCENE_DELAY_SIGMA` sets the sigma of normal distribution.
+
 ## [0.12.0] - 2021-03-03
 
 ### Changed